+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- format should write zero value of boolean in settings, by [@bradrydzewski](https://github.com/bradrydzewski).
 - Support for windows server 1903, by [@bradrydzewski](https://github.com/bradrydzewski).
 
 ## [1.2.2] - 2019-07-29